+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 Date format: DD/MM/YYYY
 
-## [3.3.1] - [20/10/2021]
+## [3.4.0] - Flexibility - [20/10/2021]
 
 - `ProgressRing` now spins on the correct direction ([#83](https://github.com/bdlukaa/fluent_ui/issues/83))
 - Added the `backwards` property to `ProgressRing`
 - `FluentApp.builder` now works as expected ([#84](https://github.com/bdlukaa/fluent_ui/issues/84))
+- Implemented `NavigationPane.customPane`, which now gives you the ability to create custom panes for `NavigationView`
 
 ## [3.3.0] - [12/10/2021]